+++ nixos/modules/features/zfs.nix
@@ -35,5 +35,27 @@ in
     ########################################################################
 
     environment.systemPackages = [ pkgs.sanoid ];
+
+    ########################################################################
+    ##                           Shell aliases                            ##
+    ########################################################################
+
+    environment.shellAliases = {
+      disable-zfs-snapshots = ''
+        sudo systemctl stop zfs-snapshot-frequent.timer && \
+        sudo systemctl stop zfs-snapshot-hourly.timer && \
+        sudo systemctl stop zfs-snapshot-daily.timer && \
+        sudo systemctl stop zfs-snapshot-weekly.timer && \
+        sudo systemctl stop zfs-snapshot-monthly.timer
+      '';
+
+      enable-zfs-snapshots = ''
+        sudo systemctl start zfs-snapshot-frequent.timer && \
+        sudo systemctl start zfs-snapshot-hourly.timer && \
+        sudo systemctl start zfs-snapshot-daily.timer && \
+        sudo systemctl start zfs-snapshot-weekly.timer && \
+        sudo systemctl start zfs-snapshot-monthly.timer
+      '';
+    };
   };
 }